requirements.txt
Flask==2.3.2
Flask-SocketIO==5.3.0
qrcode==7.3.1
Pillow==9.4.0
numpy==1.24.3
opencv-python==4.5.3.56
pyzbar==0.1.8
python-socketio==4.6.1
eventlet==0.33.3